+++ src/lib/components/room/TaskList/index.tsx
@@ -1,3 +1,4 @@
+import type { TabProps } from "@chakra-ui/react";
 import {
   Box,
   Button,
@@ -25,37 +26,49 @@ import { useForm } from "react-hook-form";
 import { GoPlus } from "react-icons/go";
 import { RiInformationLine } from "react-icons/ri";
 import { ReactSortable } from "react-sortablejs";
+import shallow from "zustand/shallow";
 
 import AutoResizeTextarea from "lib/components/shared/AutoResizeTextarea";
-import SpokerButton from "lib/components/shared/SpokerButton";
 import SpokerModalWrapper from "lib/components/shared/SpokerModalWrapper";
 import SpokerWrapperGrid from "lib/components/shared/SpokerWrapperGrid";
+import { emptyRoom } from "lib/constants/emptyRoom";
+import { useUserRole } from "lib/hooks/useUserRole";
 import { editQueueItem } from "lib/services/firebase/room/update/queue/edit";
 import { removeQueueItem } from "lib/services/firebase/room/update/queue/remove";
 import { swapSelectedQueueWithCurrent } from "lib/services/firebase/room/update/queue/swap";
 import { rewriteQueue } from "lib/services/firebase/room/update/rewriteQueue";
-import type { RoomInstance, Task } from "lib/types/RawDB";
+import { useRoomStore } from "lib/stores/room";
+import type { Task } from "lib/types/RawDB";
 
 import { submitStoryFormValidationSchema } from "./constants";
 import TaskItem from "./TaskItem";
 import type { EditStoryForm, SortableTaskItem, SubmitStoryForm } from "./types";
-
-type TaskListProps = {
-  roomData: RoomInstance;
-  showVote: boolean;
-  isOwner: boolean;
-};
 
 const initialFormValue: SubmitStoryForm = {
   name: "",
   description: "",
 };
 
-const TaskList = ({ roomData, showVote, isOwner }: TaskListProps) => {
+const tabStyle: TabProps = {
+  fontSize: {
+    base: "xs",
+    md: "md",
+  },
+};
+
+const TaskList = () => {
   const router = useRouter();
   const {
     query: { id },
   } = router;
+  const { roomData, showVote } = useRoomStore(
+    (state) => ({
+      roomData: state.roomData ?? emptyRoom,
+      showVote: state.showVote,
+    }),
+    shallow
+  );
+  const { isOwner } = useUserRole();
   const wrapperBackgroundColor = useColorModeValue("gray.50", "gray.900");
   const {
     isOpen: isOpenAddStory,
@@ -247,9 +260,9 @@ const TaskList = ({ roomData, showVote, isOwner }: TaskListProps) => {
           variant="soft-rounded"
         >
           <TabList alignItems="center">
-            {isOwner && <Tab>{activeStoriesTabText}</Tab>}
-            <Tab>{completedTabText}</Tab>
-            <Tab>{allTabText}</Tab>
+            {isOwner && <Tab {...tabStyle}>{activeStoriesTabText}</Tab>}
+            <Tab {...tabStyle}>{completedTabText}</Tab>
+            <Tab {...tabStyle}>{allTabText}</Tab>
             {isOwner && (
               <Button
                 marginLeft="auto"
@@ -342,17 +355,17 @@ const TaskList = ({ roomData, showVote, isOwner }: TaskListProps) => {
         }
         footer={
           <Flex gridGap={2}>
-            <SpokerButton onClick={onCloseAddStory} disabled={isBusy}>
+            <Button onClick={onCloseAddStory} disabled={isBusy}>
               Cancel
-            </SpokerButton>
-            <SpokerButton
+            </Button>
+            <Button
               colorScheme="teal"
               type="submit"
               isDisabled={!isValid || isBusy}
               isLoading={isBusy}
             >
               Add
-            </SpokerButton>
+            </Button>
           </Flex>
         }
       />
@@ -382,17 +395,17 @@ const TaskList = ({ roomData, showVote, isOwner }: TaskListProps) => {
         }
         footer={
           <Flex gridGap={2}>
-            <SpokerButton onClick={closeEditStory} disabled={isBusy}>
+            <Button onClick={closeEditStory} disabled={isBusy}>
               Cancel
-            </SpokerButton>
-            <SpokerButton
+            </Button>
+            <Button
               type="submit"
               colorScheme="blue"
               isDisabled={!isEditStoryValid || !isEditStoryDirty || isBusy}
               isLoading={isBusy}
             >
               Save
-            </SpokerButton>
+            </Button>
           </Flex>
         }
       />
@@ -411,15 +424,15 @@ const TaskList = ({ roomData, showVote, isOwner }: TaskListProps) => {
         }
         footer={
           <Flex gridGap={2}>
-            <SpokerButton onClick={closeRemoveStory}>Cancel</SpokerButton>
-            <SpokerButton
+            <Button onClick={closeRemoveStory}>Cancel</Button>
+            <Button
               colorScheme="red"
               onClick={processRemoveStory}
               isLoading={isBusy}
               isDisabled={isBusy}
             >
               Yes, Remove
-            </SpokerButton>
+            </Button>
           </Flex>
         }
       />