+++ src/lib/components/room/VoteWrapper.tsx
@@ -5,22 +5,27 @@ import {
   useColorModeValue,
   useRadioGroup,
 } from "@chakra-ui/react";
-import type { User } from "firebase/auth";
 import { useRouter } from "next/router";
+import shallow from "zustand/shallow";
 
 import SpokerRadioBox from "lib/components/shared/SpokerRadioBox";
 import SpokerWrapperGrid from "lib/components/shared/SpokerWrapperGrid";
+import { useUserRole } from "lib/hooks/useUserRole";
 import { updatePoint } from "lib/services/firebase/room/update/point";
-import type { RoomInstance } from "lib/types/RawDB";
+import { useAuth } from "lib/stores/auth";
+import { useRoomStore } from "lib/stores/room";
 import { pointOptions } from "lib/types/room";
 
-type VoteWrapperProps = {
-  roomData: RoomInstance;
-  currentUser: User;
-  showVote: boolean;
-};
-
-const VoteWrapper = ({ roomData, currentUser, showVote }: VoteWrapperProps) => {
+const VoteWrapper = () => {
+  const currentUser = useAuth((state) => state.currentUser);
+  const { roomData, showVote } = useRoomStore(
+    (state) => ({
+      roomData: state.roomData,
+      showVote: state.showVote,
+    }),
+    shallow
+  );
+  const { isOwner, isParticipant } = useUserRole();
   const wrapperBackgroundColor = useColorModeValue("orange.50", "gray.600");
   const router = useRouter();
   const {
@@ -28,19 +33,25 @@ const VoteWrapper = ({ roomData, currentUser, showVote }: VoteWrapperProps) => {
   } = router;
 
   const handleUpdatePoint = async (point: number) => {
-    if (currentUser && !(roomData.config.isFreezeAfterVote && showVote)) {
+    if (currentUser && !(roomData?.config.isFreezeAfterVote && showVote)) {
       await updatePoint({ uid: currentUser.uid, point, roomId: id as string });
     }
   };
 
   const { getRootProps, getRadioProps } = useRadioGroup({
     name: "vote",
-    value: currentUser && String(roomData?.users?.[currentUser.uid]?.point),
+    value: currentUser
+      ? String(roomData?.users?.[currentUser.uid]?.point)
+      : undefined,
     onChange: (value) => {
       handleUpdatePoint(Number(value));
     },
   });
   const voteOptionGroup = getRootProps();
+
+  if (!isOwner && !isParticipant) {
+    return null;
+  }
 
   return (
     <SpokerWrapperGrid gap={4} backgroundColor={wrapperBackgroundColor}>