+++ src/lib/components/room/header/index.tsx
@@ -10,12 +10,15 @@ import {
 import * as React from "react";
 
 import SpokerWrapperGrid from "lib/components/shared/SpokerWrapperGrid";
+import { useUserRole } from "lib/hooks/useUserRole";
+import { useRoomStore } from "lib/stores/room";
 
 import EditableFields from "./components/EditableFields";
 import { useRoomHeader } from "./hooks";
-import type { RoomHeaderProps } from "./types";
 
-const RoomHeader = ({ roomData, isOwner }: RoomHeaderProps) => {
+const RoomHeader = () => {
+  const roomData = useRoomStore((state) => state.roomData);
+  const { isOwner } = useUserRole();
   const wrapperBackgroundColor = useColorModeValue("teal.50", "teal.600");
   const { name, description, handleUpdateTask } = useRoomHeader({
     roomData,
@@ -23,7 +26,6 @@ const RoomHeader = ({ roomData, isOwner }: RoomHeaderProps) => {
   });
 
   const content = React.useMemo(() => {
-    const task = roomData?.task;
     if (isOwner) {
       return (
         <EditableFields
@@ -32,6 +34,10 @@ const RoomHeader = ({ roomData, isOwner }: RoomHeaderProps) => {
           handleUpdateTask={handleUpdateTask}
         />
       );
+    }
+    const task = roomData?.task;
+    if (!task) {
+      return null;
     }
 
     return (