+++ src/lib/components/room/header/types.ts
@@ -1,6 +1,5 @@
 import type { RoomInstance } from "lib/types/RawDB";
 
 export type RoomHeaderProps = {
-  roomData: RoomInstance;
-  isOwner: boolean;
+  roomData?: RoomInstance;
 };