+++ src/lib/pages/room/index.tsx
@@ -1,5 +1,6 @@
 import { Box, Grid } from "@chakra-ui/react";
 import { NextSeo } from "next-seo";
+import shallow from "zustand/shallow";
 
 import ControllerWrapper from "lib/components/room/ControllerWrapper";
 import CurrentVotesWrapper from "lib/components/room/CurrentVotesWrapper";
@@ -7,25 +8,25 @@ import RoomHeader from "lib/components/room/header";
 import TaskList from "lib/components/room/TaskList";
 import VoteWrapper from "lib/components/room/VoteWrapper";
 import SpokerLoading from "lib/components/shared/SpokerLoading";
-
-import { useRoom } from "./hooks";
+import { useRoomListener } from "lib/hooks/useRoomListener";
+import { useVoteListener } from "lib/hooks/useVoteListener";
+import { useAuth } from "lib/stores/auth";
+import { useRoomStore } from "lib/stores/room";
 
 const RoomContainer = () => {
-  const {
-    currentUser,
-    busy,
-    showVote,
-    roomData,
-    users,
-    averagePoint,
-    highestPoint,
-    isParticipant,
-    isObservant,
-    isOwner,
-    handleFinishVote,
-  } = useRoom();
+  const currentUser = useAuth((state) => state.currentUser);
+  const { isBusy, roomData } = useRoomStore(
+    (state) => ({
+      isBusy: state.isBusy,
+      roomData: state.roomData,
+    }),
+    shallow
+  );
 
-  if (busy) {
+  useRoomListener();
+  useVoteListener();
+
+  if (isBusy) {
     return <SpokerLoading />;
   }
 
@@ -39,33 +40,17 @@ const RoomContainer = () => {
           alignItems="start"
         >
           <Grid gap={6}>
-            <RoomHeader roomData={roomData} isOwner={isOwner} />
-            {(isOwner || isParticipant) && (
-              <VoteWrapper
-                roomData={roomData}
-                currentUser={currentUser}
-                showVote={showVote}
-              />
-            )}
+            <RoomHeader />
+            <VoteWrapper />
           </Grid>
 
           <Grid gap={6}>
-            <CurrentVotesWrapper
-              isOwner={isOwner}
-              isObservant={isObservant}
-              roomData={roomData}
-              showVote={showVote}
-              averagePoint={averagePoint}
-              highestPoint={highestPoint}
-              users={users}
-              currentUser={currentUser}
-              onFinishVote={handleFinishVote}
-            />
-            <ControllerWrapper users={users} isResetEnabled={isOwner} />
+            <CurrentVotesWrapper />
+            <ControllerWrapper />
           </Grid>
         </Grid>
 
-        <TaskList isOwner={isOwner} roomData={roomData} showVote={showVote} />
+        <TaskList />
       </Grid>
     );
   }